Write nicer version also to log.
